update
1.	Decode Method
a.	Old: interpret 4 digits
b.	New: interpret 6 digits where 1st 3 number are opcode, and the last 3 digit is the operand

2.	Opcode execution
a.	Verified the opcode in executing the 6-digit format
b.	Verified if invalid opcode and halt if invalid opcode

3.	Overflow
a.	Old: limits to 4 digits 9999 and -9999
b.	New:  limits to 6 digits 999999 and -999999

4.	Save feature
a.	Old: save button won’t allow user to rename file
b.	New: can custom filename when save

5.	UI adjustment
a.	Correctly display the 6 digits for instruction and memory
b.	Support 6 digit, ensure the new instruction format aligned to user input